Rename futureInstaller to init
